Added tantalus stage 1 states to config file (and a bunch of miscl bugfixes)

+++ avionics/common/include/states/main_descent.h
@@ -9,23 +9,23 @@ namespace State {
 
     class MainDescent : IState {
     public:
-        /*
-        * @param float LANDED_TIME_INTERVAL time between each landed check, in seconds
-        * @param uint8_t LANDED_CHECKS number of checks to perform
-        * @param float LANDED_VELOCITY maximum velocity to be considered as landed
+        /**
+        * @param LANDED_TIME_INTERVAL time between each landed check, in milliseconds
+        * @param LANDED_CHECKS number of checks to perform
+        * @param LANDED_VELOCITY maximum velocity to be considered as landed
         */
-        MainDescent(float LANDED_TIME_INTERVAL, uint8_t LANDED_CHECKS, float LANDED_VELOCITY) :
+        MainDescent(uint32_t LANDED_TIME_INTERVAL, uint8_t LANDED_CHECKS, float LANDED_VELOCITY) :
             M_LANDED_TIME_INTERVAL(LANDED_TIME_INTERVAL),
             M_LANDED_CHECKS(LANDED_CHECKS),
             M_LANDED_VELOCITY(LANDED_VELOCITY) {}
 
-        /*
+        /**
         * @brief Return the assigned enumeration code.
         * @return Enumeration code.
         */
         StateId getStateEnum(void){ return StateId::MAIN_DESCENT; }
 
-        /*
+        /**
         * @brief Return the next state, based on input data (mostly from filtered sensor data)
         * @return State enumeration code, to be passed into the std::map between codes and used states. Note that the returned code may be the same state.
         */
